add anonymous inner class test
An anonymous inner class cannot have constructors because the name of a
constructor must be the same as the name of a class, and the class has no
name. Instead, the construction parameters are given to the superclass
constructor. In particular, whenever an inner class implements an
interface, it cannot have any construction parameters. Nevertheless, you
must supply a set of parenthesis as in
new InterfaceType()
{
    methods and data
}
You have to look carefully to see the difference betweeen the construction
of a new object of a class and the construction of an object of an
anonymous innner class extending that class.
Person queen = new Person("Mary");
// a person object
Person count = new Person("Dracula") {...}
// an object of an inner class extending Person

if the closing parenthesis of the construction parameter list is followed
by an opening brace, then an anonymous inner class is being defined.