add example of static inner class
Occasionally, we want to use an inner class simply to hide one class
inside another, but we don't need the inner class to have a reference to
the outer class object. We can suppress the generation of that reference
by declaring the inner class static.

A static inner class is exactly like any other inner class, except that
an object of a static inner class does not have a reference to the outer
class object that generated it. In our example, we must use a static
inner class because the inner class object is constructed inside a
static method.

public static Pair minmax(double[]  d)
{
    ...
    return new Pair(min, max);
}

Had the Pair class not been declared as static, the compiler would have
complained that there was no implicit object of type ArrayAlt available
to initialize the inner class object.