demonstrates the use of proxy class
The proxy class can create a brand-new classes at runtime. Such a proxy
class implements the interfaces that you specify.In particular, the
proxy class has the following methods:
- all methods required by the specified interfaces and
- all methods defined in the Object class (toString, equals, and so on)
However, you cannot define new code, for these methods at runtime.
Instead, you must supply an invocation handler. An invocation handler is
an object of type any class that implementsthe InvocationHandler
interface. That interface has a single method:
    Object invoke(Object proxy, Method method, Object[] args)
    
Whenever a method is called on the proxy object, the invoke method of
the invocation handler gets called, with the Method object and
parameters of the original call. The invocation handler must then figure
out how to handle the call.

To create a proxy object, you use the newProxyInstance method of the
proxy class. The method has three parameters:
- A class loader
- An array of class objects, one for each interface to be implemented.
- An invocation handler