use toolkit to get the screensize
toolkit is used to get the screen size of the device on which
application is currently executed. The application screen size is then
resized to fit relative to the device screen size. The platform is let
to decide where to place the window. An icon is also added to the window
using toolkit.